+++ js/components/generate/generate.jsx
@@ -1,0 +1,4 @@
+import ReactDOM from "react-dom"
+import GenerateForm from "./GenerateForm.js"
+
+ReactDOM.render(<GenerateForm />, document.getElementById("generateRoot"))