+++ webpack.config.js
@@ -4,7 +4,7 @@ const webpack = require("webpack")
 
 const presets = [
   ["@babel/preset-env", { useBuiltIns: "usage", corejs: 3 }],
-  ["@babel/preset-react"],
+  ["@babel/preset-react", {runtime: "automatic"}],
 ]
 
 const isDevelopment = process.env.NODE_ENV !== "production"
@@ -17,7 +17,7 @@ module.exports = {
     "js/graph/app": "./js/components/graph/main.js",
     "js/post/app": "./js/components/post/post.js.jsx",
     "js/draw/app": "./js/components/draw/main.js",
-    "js/generate/app": "./js/components/generate/generate.js",
+    "js/generate/app": "./js/components/generate/generate.jsx",
     "style/app": "./style/app.js",
   },
   output: {